refactor: 쓸모없는 코드 삭제
